Обновление окружения для celery
