+++ api/wsgi.py
@@ -1,0 +1,3 @@
+from api import create_application
+
+application = create_application()
