--- site_parser/wsgi.py
@@ -1,3 +1,0 @@
-from site_parser import create_application
-
-application = create_application()
